Impelement "Share to LOVE" support.
This allows you to share either .love game to launch as game (for the
non-fused APKs and instance of LOVE isn't running) or any file that
gets passed as love.filedropped.

Caveat: Ensure the intended game is running before sending files to LOVE app.
Otherwise, LOVE will attempt to open it as *.love file, which is may not what
you want.